# docs/swagger.yaml
basePath: /api
definitions:
  models.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  models.ApiResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/models.ApiError'
      success:
        type: boolean
    type: object
  models.Category:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    required:
    - name
    type: object
  models.CategoryCreateRequest:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  models.CategoryListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  models.CategoryUpdateRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: bruno.malagoli@example.com
    name: Bsmart API Support
  description: RESTful API con soporte de WebSocket en tiempo real para gestión de
    productos y categorías. Sistema completo con autenticación JWT, control de acceso
    basado en roles, búsqueda full-text y seguimiento automático de historial de productos.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bsmart Backend API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Obtiene la lista completa de categorías con búsqueda opcional por
        nombre
      parameters:
      - description: Término de búsqueda en nombre de categoría
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de categorías
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.CategoryListResponse'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Listar categorías
      tags:
      - categorías
    post:
      consumes:
      - application/json
      description: Crea una nueva categoría en el sistema. Requiere rol de administrador.
        Emite evento WebSocket 'category:created'.
      parameters:
      - description: Datos de la categoría
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Categoría creada exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Category'
              type: object
        "400":
          description: Datos de entrada inválidos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "403":
          description: Requiere rol de administrador
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Crear nueva categoría
      tags:
      - categorías
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una categoría del sistema. Requiere rol de administrador.
        Emite evento WebSocket 'category:deleted'.
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Categoría eliminada exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: ID inválido
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "403":
          description: Requiere rol de administrador
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "404":
          description: Categoría no encontrada
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Eliminar categoría
      tags:
      - categorías
    get:
      consumes:
      - application/json
      description: Obtiene la información completa de una categoría específica
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Detalle de la categoría
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Category'
              type: object
        "400":
          description: ID inválido
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "404":
          description: Categoría no encontrada
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Obtener detalle de categoría
      tags:
      - categorías
    put:
      consumes:
      - application/json
      description: Actualiza una categoría existente. Requiere rol de administrador.
        Los campos no enviados no se modifican. Emite evento WebSocket 'category:updated'.
      parameters:
      - description: ID de la categoría
        in: path
        name: id
        required: true
        type: integer
      - description: Datos a actualizar (campos opcionales)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CategoryUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Categoría actualizada exitosamente
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Category'
              type: object
        "400":
          description: ID inválido o datos de entrada inválidos
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "401":
          description: No autenticado
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "403":
          description: Requiere rol de administrador
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "404":
          description: Categoría no encontrada
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
        "500":
          description: Error interno del servidor
          schema:
            allOf:
            - $ref: '#/definitions/models.ApiResponse'
            - properties:
                error:
                  $ref: '#/definitions/models.ApiError'
              type: object
      security:
      - BearerAuth: []
      summary: Actualizar categoría
      tags:
      - categorías
...
securityDefinitions:
  BearerAuth:
    description: 'Autenticación JWT. Formato: "Bearer {token}". Obtén un token desde
      /api/auth/login o /api/auth/register'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"